Remagi

+++ src/MCModels/Theater.java
@@ -10,6 +10,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import static java.sql.Types.NULL;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -28,18 +29,14 @@ public class Theater {
     private String name;
     private int rowCount;
     private int colCount;
-
     private int status;
-    private int Row;
-    private int Collum;
 
-    public Theater(int id, int cinemaID, int Number, int status, int row, int collum) {
+    public Theater(int id, int cinemaID, int managerID, String name, int status) {
         this.id = id;
         this.cinemaID = cinemaID;
-        this.Number = Number;
+        this.managerID = managerID;
+        this.name = name;
         this.status = status;
-        this.Row = row;
-        this.Collum = collum;
     }
 
     public Theater() {
@@ -53,8 +50,8 @@ public class Theater {
         return cinemaID;
     }
 
-    public int getNumber() {
-        return Number;
+    public int getManagerID() {
+        return managerID;
     }
 
     public String getName() {
@@ -64,15 +61,6 @@ public class Theater {
     public int getStatus() {
         return status;
     }
-
-    public int getRow() {
-        return Row;
-    }
-
-    public int getCollum() {
-        return Collum;
-    }
-    
 
     public void setId(int id) {
         this.id = id;
@@ -88,7 +76,6 @@ public class Theater {
 
     public void setName(String name) {
         this.name = name;
-
     }
 
     public void setStatus(int status) {
@@ -109,86 +96,6 @@ public class Theater {
 
     public void setColCount(int colCount) {
         this.colCount = colCount;
-    }
-
-    public int Insert(Theater theater) {
-
-        try {
-
-
-            String sqlstr = "insert into theater(theaterCinema, theaterManagerID, theaterStatus) values( "
-                    + theater.getCinemaID() + ", " + theater.getManagerID() + ", " + theater.getStatus()
-                    + " )";
-
-            Database db = new Database();
-
-            db.openConnection();
-
-            ResultSet rst = db.getInsertObjectIDs(sqlstr);  
-
-            if (rst != null && rst.first()) {
-                theater.setId(rst.getInt(1));
-            }
-
-            db.close(rst);
-            //db.closeConnection();	
-            return 1;
-
-        } catch (Exception e) {
-            //e.printStackTrace();
-            System.out.println(e.getMessage());
-
-        }
-        return 0;
-    }
-
-    public int Update(Theater theater) {
-
-        int rtn = 0;
-
-
-        try {
-            String sqlstr = "update theater set " + " theaterCinemaID = " + theater.getCinemaID()
-                    + ", theaterManager = " + theater.getManagerID() + ", theaterStatus = " + theater.getStatus();
-
-            //String sqlstr =" ";
-            sqlstr += " where theaterId = " + theater.getId();
-
-            Database db = new Database();
-
-            db.openConnection();
-
-            rtn = db.execCommand(sqlstr);
-
-            //db.closeConnection();	
-        } catch (Exception e) {
-            //e.printStackTrace();
-            System.out.println(e.getMessage());
-        }
-        return rtn;
-    }
-
-    public int Delete(int ID) {
-
-        int rtn = 0;
-
-        try {
-            String sqlstr = "delete from theater ";
-
-            sqlstr += " where theaterId = " + ID;
-
-            Database db = new Database();
-
-            db.openConnection();
-
-            rtn = db.execCommand(sqlstr);
-
-            //db.closeConnection();
-        } catch (Exception e) {
-            //e.printStackTrace();
-            System.out.println(e.getMessage());
-        }
-        return rtn;
     }
 
     public ArrayList<Theater> loadTheaters() {
@@ -229,7 +136,6 @@ public class Theater {
 
         } catch (Exception ex) {
             Logger.getLogger(Theater.class.getName()).log(Level.SEVERE, null, ex);
-
         }
 
         return temp;
@@ -239,7 +145,7 @@ public class Theater {
     public int insert(Theater theater) {
         try {
             String sqlstr = "insert into theater( theaterName, theaterRowCount, theaterColCount, theaterStatus ) values( '"
-                    + theater.getName() + "', " + theater.getRowCount() + ", " + theater.getColCount() + "', 1 )";
+                    + theater.getName() + "', " + theater.getRowCount() + ", " + theater.getColCount() + ", 1 )";
             Database db = new Database();
             db.openConnection();
             ResultSet rst = db.getInsertObjectIDs(sqlstr);
@@ -275,15 +181,22 @@ public class Theater {
 
     public int delete(int id) {
         int rtn = 0;
-        try {
-            String sqlstr = "delete from theater ";
-            sqlstr += " where theaterId = " + id;
-            Database db = new Database();
-            db.openConnection();
-            rtn = db.execCommand(sqlstr);
-            db.close();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+        Theater theater = new Theater();
+        theater.setId(id);
+        theater.setRowCount(theater.select("theaterId=" + theater.getId()).get(0).getRowCount());
+        theater.setColCount(theater.select("theaterId=" + theater.getId()).get(0).getColCount());
+
+        if (deleteSeats(theater) != 0) {
+            try {
+                String sqlstr = "delete from theater ";
+                sqlstr += " where theaterId = " + id;
+                Database db = new Database();
+                db.openConnection();
+                rtn = db.execCommand(sqlstr);
+                db.close();
+            } catch (Exception e) {
+                System.out.println(e.getMessage());
+            }
         }
         return rtn;
     }
@@ -333,6 +246,31 @@ public class Theater {
                     seat.setColumn(j);
                     seat.setStatus(1);
                     tempSeat.insert(seat);
+                }
+            }
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return 0;
+        }
+        return 1;
+    }
+
+    public int deleteSeats(Theater theater) {
+        //Seat seat = new Seat();
+        //Seat tempSeat = new Seat();
+        try {
+            for (int i = 1; i <= theater.getRowCount(); i++) {
+                for (int j = 1; j <= theater.getColCount(); j++) {
+                    Seat seat = new Seat();
+                    Seat tempSeat = new Seat();
+
+                    seat.setTheaterId(theater.getId());
+                    System.out.println(seat.getTheaterId());
+                    seat.setRow(i);
+                    seat.setColumn(j);
+                    int seatId = (tempSeat.select("theaterId=" + seat.getTheaterId() + ",seatRow=" + seat.getRow() + ",seatCol=" + seat.getColumn())).get(0).getId();
+
+                    tempSeat.delete(seatId);
                 }
             }
         } catch (Exception e) {
